+++ webpack-babel-react/client/.eslintrc
