+++ webpack-babel-react/client/components/App/App.jsx
