+++ webpack-babel-react/client/config/setup.js
