+++ webpack-babel-react/client/index.jsx
@@ -1,11 +1,14 @@
 import { Router, browserHistory } from 'react-router';
+import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 
-import store from './store';
+import store from './store/store';
 import routes from './routes';
+import './setup';
 
-export default (
+render(
   <Provider store={store}>
     <Router history={browserHistory}>{routes}</Router>
-  </Provider>
+  </Provider>,
+  document.getElementById('root')
 );
