+++ webpack-babel-react/client/reducers/reducer.js
