+++ webpack-babel-react/client/routes.jsx
