+++ webpack-babel-react/client/store/store.js
