--- webpack-babel-react/ui/entry.js
@@ -1,6 +1,0 @@
-import ReactDOM from 'react-dom';
-
-import './setup';
-import App from './app';
-
-ReactDOM.render(App, document.getElementById('root'));
